backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.40.0
httpx>=0.27.0
pydantic>=2.7.0
starlette>=0.49.1
python-dotenv>=1.0.0
asyncpg>=0.29.0
python-multipart>=0.0.22
